Introduce a common `checkPlaceholders()` implementation
