+++ model/src/main/kotlin/io/spine/validation/required/RequiredOption.kt
@@ -55,12 +55,12 @@ import io.spine.validation.IF_MISSING
 import io.spine.validation.OPTION_NAME
 import io.spine.validation.REQUIRED
 import io.spine.validation.checkBothApplied
+import io.spine.validation.checkPlaceholders
 import io.spine.validation.defaultErrorMessage
 import io.spine.validation.event.IfMissingOptionDiscovered
 import io.spine.validation.event.RequiredFieldDiscovered
 import io.spine.validation.event.ifMissingOptionDiscovered
 import io.spine.validation.event.requiredFieldDiscovered
-import io.spine.validation.missingPlaceholders
 import io.spine.validation.required.RequiredFieldSupport.isSupported
 
 /**
@@ -134,7 +134,7 @@ internal class IfMissingPolicy : Policy<FieldOptionDiscovered>() {
 
         val option = event.option.unpack<IfMissingOption>()
         val message = option.errorMsg
-        checkPlaceholders(message, field, file)
+        message.checkPlaceholders(SUPPORTED_PLACEHOLDERS, field, file, IF_MISSING)
 
         return ifMissingOptionDiscovered {
             id = field.ref
@@ -149,14 +149,5 @@ private fun checkFieldType(field: Field, file: File) =
                 " by the `($REQUIRED)` option. Supported field types: messages, enums," +
                 " strings, bytes, repeated, and maps."
     }
-
-private fun checkPlaceholders(template: String, field: Field, file: File) {
-    val missing = missingPlaceholders(template, SUPPORTED_PLACEHOLDERS)
-    Compilation.check(missing.isEmpty(), file, field.span) {
-        "The `${field.qualifiedName}` field specifies an error message using the `($IF_MISSING)`" +
-                " option with unsupported placeholders: `$missing`. Supported placeholders are" +
-                " the following: `${SUPPORTED_PLACEHOLDERS.map { it.value }}`."
-    }
-}
 
 private val SUPPORTED_PLACEHOLDERS = setOf(FIELD_PATH, FIELD_TYPE, PARENT_TYPE)
